Move search/replace functions from kraken-build to lib
